CaseManager.Configuration (Case Manager v0.1.0)

Domain that controls everything related to app configuration.

Functions

add_setting(key, value, params \\ nil, opts \\ nil)

Calls the set_setting action on CaseManager.Configuration.Setting.

Arguments

key

value

Options

:upsert? (boolean/0) - If a conflict is found based on the primary key, the record is updated in the database (requires upsert support) The default value is false.

:return_skipped_upsert? (boolean/0) - If true, and a record was not upserted because its filter prevented the upsert, the original record (which was not upserted) will be returned. The default value is false.

:upsert_identity (atom/0) - The identity to use when detecting conflicts for upsert?, e.g. upsert_identity: :full_name. By default, the primary key is used. Has no effect if upsert?: true is not provided

:upsert_fields - The fields to upsert. If not set, the action's upsert_fields is used, and if that is not set, then any fields not being set to defaults are written.

:upsert_condition (term/0) - An expression to check if the record should be updated when there's a conflict.

:timeout (timeout/0) - A positive integer, or :infinity. If none is provided, the timeout configured on the domain is used.

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:action (term/0) - The action to use, either an Action struct or the name of the action

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user. Valid values are true, false, nil

:context (map/0) - Context to set on the query, changeset, or input

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:return_notifications? (boolean/0) - Use this if you're running ash actions in your own transaction and you want to manually handle sending notifications.
If a transaction is ongoing, and this is false, notifications will be discarded, otherwise the return value is {:ok, result, notifications} (or {:ok, notifications})
To send notifications later, use Ash.Notifier.notify(notifications). It sends any notifications that can be sent, and returns the rest. The default value is false.

:rollback_on_error? (boolean/0) - Whether or not to rollback the transaction on error, if the resource is in a transaction.
If the action has transaction? false this option has no effect. If an error is returned from the data layer and the resource is in a transaction, the transaction is always rolled back, regardless. The default value is true.

:notification_metadata (term/0) - Metadata to be merged into the metadata field for all notifications sent from this operation. The default value is %{}.

:skip_unknown_inputs - A list of inputs that, if provided, will be ignored if they are not recognized by the action. Use :* to indicate all unknown keys.

:load (term/0) - A load statement to add onto the changeset

:changeset (term/0) - A changeset to seed the action with.

:bulk_options (keyword/0) - Options passed to Ash.bulk_create, if a list or stream of inputs is provided.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:after_action (function of arity 2) - An after_action hook to be added to each processed changeset

:read_action (atom/0) - The action to use when building the read query.

:assume_casted? (boolean/0) - Whether or not to cast attributes and arguments as input. This is an optimization for cases where the input is already casted and/or not in need of casting The default value is false.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:authorize_query_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_changeset_with - If set to :error, instead of filtering unauthorized changes, unauthorized changes will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. The default value is :filter.

:sorted? (boolean/0) - Whether or not to sort results by their input position, in cases where return_records?: true was provided. The default value is false.

:return_records? (boolean/0) - Whether or not to return all of the records that were inserted. Defaults to false to account for large inserts. The default value is false.

:return_errors? (boolean/0) - Whether or not to return all of the errors that occur. Defaults to false to account for large inserts. The default value is false.

:batch_size (pos_integer/0) - The number of records to include in each batch. Defaults to the default_limit or max_page_size of the action, or 100.

:return_stream? (boolean/0) - If set to true, instead of an Ash.BulkResult, a mixed stream is returned. Potential elements: {:notification, notification} - if return_notifications? is set to true {:ok, record} - if return_records? is set to true {:error, error} - an error that occurred. May be changeset or an individual error. The default value is false.

:return_nothing? (boolean/0) - Mutes warnings about returning nothing. Only relevant if return_stream? is set to true and all other return_*? options are set to false. The default value is false.

:stop_on_error? (boolean/0) - If true, the first encountered error will stop the action and be returned. Otherwise, errors will be skipped. The default value is false.

:notify? (boolean/0) - Whether or not to generate any notifications. If this is set to true then the data layer must return the results from each batch. This may be intensive for large bulk actions. Notifications will be automatically sent unless return_notifications? is set to true. The default value is false.

:transaction - Whether or not to wrap the entire execution in a transaction, each batch, or not at all. Keep in mind: before_transaction and after_transaction hooks attached to changesets will have to be run inside the transaction if you choose transaction: :all. The default value is :batch.

:max_concurrency (non_neg_integer/0) - If set to a value greater than 0, up to that many tasks will be started to run batches asynchronously The default value is 0.

add_setting!(key, value, params \\ nil, opts \\ nil)

Calls the set_setting action on CaseManager.Configuration.Setting. Raises any errors instead of returning them

Arguments

key

value

Options

:upsert? (boolean/0) - If a conflict is found based on the primary key, the record is updated in the database (requires upsert support) The default value is false.

:return_skipped_upsert? (boolean/0) - If true, and a record was not upserted because its filter prevented the upsert, the original record (which was not upserted) will be returned. The default value is false.

:upsert_identity (atom/0) - The identity to use when detecting conflicts for upsert?, e.g. upsert_identity: :full_name. By default, the primary key is used. Has no effect if upsert?: true is not provided

:upsert_fields - The fields to upsert. If not set, the action's upsert_fields is used, and if that is not set, then any fields not being set to defaults are written.

:upsert_condition (term/0) - An expression to check if the record should be updated when there's a conflict.

:timeout (timeout/0) - A positive integer, or :infinity. If none is provided, the timeout configured on the domain is used.

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:action (term/0) - The action to use, either an Action struct or the name of the action

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user. Valid values are true, false, nil

:context (map/0) - Context to set on the query, changeset, or input

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:return_notifications? (boolean/0) - Use this if you're running ash actions in your own transaction and you want to manually handle sending notifications.
If a transaction is ongoing, and this is false, notifications will be discarded, otherwise the return value is {:ok, result, notifications} (or {:ok, notifications})
To send notifications later, use Ash.Notifier.notify(notifications). It sends any notifications that can be sent, and returns the rest. The default value is false.

:rollback_on_error? (boolean/0) - Whether or not to rollback the transaction on error, if the resource is in a transaction.
If the action has transaction? false this option has no effect. If an error is returned from the data layer and the resource is in a transaction, the transaction is always rolled back, regardless. The default value is true.

:notification_metadata (term/0) - Metadata to be merged into the metadata field for all notifications sent from this operation. The default value is %{}.

:skip_unknown_inputs - A list of inputs that, if provided, will be ignored if they are not recognized by the action. Use :* to indicate all unknown keys.

:load (term/0) - A load statement to add onto the changeset

:changeset (term/0) - A changeset to seed the action with.

:bulk_options (keyword/0) - Options passed to Ash.bulk_create, if a list or stream of inputs is provided.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:after_action (function of arity 2) - An after_action hook to be added to each processed changeset

:read_action (atom/0) - The action to use when building the read query.

:assume_casted? (boolean/0) - Whether or not to cast attributes and arguments as input. This is an optimization for cases where the input is already casted and/or not in need of casting The default value is false.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:authorize_query_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_changeset_with - If set to :error, instead of filtering unauthorized changes, unauthorized changes will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. The default value is :filter.

:sorted? (boolean/0) - Whether or not to sort results by their input position, in cases where return_records?: true was provided. The default value is false.

:return_records? (boolean/0) - Whether or not to return all of the records that were inserted. Defaults to false to account for large inserts. The default value is false.

:return_errors? (boolean/0) - Whether or not to return all of the errors that occur. Defaults to false to account for large inserts. The default value is false.

:batch_size (pos_integer/0) - The number of records to include in each batch. Defaults to the default_limit or max_page_size of the action, or 100.

:return_stream? (boolean/0) - If set to true, instead of an Ash.BulkResult, a mixed stream is returned. Potential elements: {:notification, notification} - if return_notifications? is set to true {:ok, record} - if return_records? is set to true {:error, error} - an error that occurred. May be changeset or an individual error. The default value is false.

:return_nothing? (boolean/0) - Mutes warnings about returning nothing. Only relevant if return_stream? is set to true and all other return_*? options are set to false. The default value is false.

:stop_on_error? (boolean/0) - If true, the first encountered error will stop the action and be returned. Otherwise, errors will be skipped. The default value is false.

:notify? (boolean/0) - Whether or not to generate any notifications. If this is set to true then the data layer must return the results from each batch. This may be intensive for large bulk actions. Notifications will be automatically sent unless return_notifications? is set to true. The default value is false.

:transaction - Whether or not to wrap the entire execution in a transaction, each batch, or not at all. Keep in mind: before_transaction and after_transaction hooks attached to changesets will have to be run inside the transaction if you choose transaction: :all. The default value is :batch.

:max_concurrency (non_neg_integer/0) - If set to a value greater than 0, up to that many tasks will be started to run batches asynchronously The default value is 0.

can_add_setting(actor, key, value, params_or_opts \\ %{}, opts \\ [])

Runs authorization checks for CaseManager.Configuration.Setting.set_setting See Ash.can/3 for more information

Options

:maybe_is (term/0) - If the actor may be able to perform the action, what value should be returned. The default value is :maybe.

:filter_with - If set to :error, the query will raise an error on a match. If set to :filter the query will filter out unauthorized access. Valid values are :filter, :error The default value is :filter.

:validate? (boolean/0) - Whether or not to treat an invalid action as a non-allowed action. The default value is false.

:reuse_values? (boolean/0) - Whether or not loaded data like aggregates, calculations and relationships should be checked in memory if possible, instead of querying. No effect if pre_flight? is false. The default value is false.

:pre_flight? (boolean/0) - Whether or not this is a pre_flight check (which may perform optimized in-memory checks) or the final proper check. The default value is true.

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol. Will overwrite any actor, tenant or context provided. See Ash.Context for more.

:run_queries? (boolean/0) - Whether or not to run queries. If set to true, :maybe will not be returned. The default value is true.

:data - The record or records specifically attempting to be acted upon.

:tenant (value that implements the Ash.ToTenant protocol) - The tenant to use for authorization

:alter_source? (boolean/0) - If set to true, the source being authorized is returned so it can be run. The default value is false.

:base_query (term/0) - A base query on which to apply an generated filters

:no_check? (boolean/0) - Whether or not authorization must pass at the strict/filter step, or if post-checks are allowed to be run The default value is false.

:on_must_pass_strict_check (term/0) - Override the value returned when no_check? is true but a check must be run.

:atomic_changeset (term/0) - A base query on which to apply an generated filters

:return_forbidden_error? (boolean/0) - Whether or not to return a forbidden error in cases of not being authorized. The default value is false.

can_add_setting?(actor, key, value, params_or_opts \\ %{}, opts \\ [])

Runs authorization checks for CaseManager.Configuration.Setting.set_setting, returning a boolean. See Ash.can?/3 for more information

Options

:maybe_is (term/0) - If the actor may be able to perform the action, what value should be returned. The default value is true.

:filter_with - If set to :error, the query will raise an error on a match. If set to :filter the query will filter out unauthorized access. Valid values are :filter, :error The default value is :filter.

:validate? (boolean/0) - Whether or not to treat an invalid action as a non-allowed action. The default value is false.

:reuse_values? (boolean/0) - Whether or not loaded data like aggregates, calculations and relationships should be checked in memory if possible, instead of querying. No effect if pre_flight? is false. The default value is false.

:pre_flight? (boolean/0) - Whether or not this is a pre_flight check (which may perform optimized in-memory checks) or the final proper check. The default value is true.

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol. Will overwrite any actor, tenant or context provided. See Ash.Context for more.

:run_queries? (boolean/0) - Whether or not to run queries. If set to true, :maybe will not be returned. The default value is true.

:data - The record or records specifically attempting to be acted upon.

:tenant (value that implements the Ash.ToTenant protocol) - The tenant to use for authorization

:alter_source? (boolean/0) - If set to true, the source being authorized is returned so it can be run. The default value is false.

:base_query (term/0) - A base query on which to apply an generated filters

:no_check? (boolean/0) - Whether or not authorization must pass at the strict/filter step, or if post-checks are allowed to be run The default value is false.

:on_must_pass_strict_check (term/0) - Override the value returned when no_check? is true but a check must be run.

:atomic_changeset (term/0) - A base query on which to apply an generated filters

:return_forbidden_error? (boolean/0) - Whether or not to return a forbidden error in cases of not being authorized. The default value is false.

can_upload_file_to_setting(actor, key, value, file, params_or_opts \\ %{}, opts \\ [])

Runs authorization checks for CaseManager.Configuration.Setting.upload_file See Ash.can/3 for more information

Options

:maybe_is (term/0) - If the actor may be able to perform the action, what value should be returned. The default value is :maybe.

:filter_with - If set to :error, the query will raise an error on a match. If set to :filter the query will filter out unauthorized access. Valid values are :filter, :error The default value is :filter.

:validate? (boolean/0) - Whether or not to treat an invalid action as a non-allowed action. The default value is false.

:reuse_values? (boolean/0) - Whether or not loaded data like aggregates, calculations and relationships should be checked in memory if possible, instead of querying. No effect if pre_flight? is false. The default value is false.

:pre_flight? (boolean/0) - Whether or not this is a pre_flight check (which may perform optimized in-memory checks) or the final proper check. The default value is true.

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol. Will overwrite any actor, tenant or context provided. See Ash.Context for more.

:run_queries? (boolean/0) - Whether or not to run queries. If set to true, :maybe will not be returned. The default value is true.

:data - The record or records specifically attempting to be acted upon.

:tenant (value that implements the Ash.ToTenant protocol) - The tenant to use for authorization

:alter_source? (boolean/0) - If set to true, the source being authorized is returned so it can be run. The default value is false.

:base_query (term/0) - A base query on which to apply an generated filters

:no_check? (boolean/0) - Whether or not authorization must pass at the strict/filter step, or if post-checks are allowed to be run The default value is false.

:on_must_pass_strict_check (term/0) - Override the value returned when no_check? is true but a check must be run.

:atomic_changeset (term/0) - A base query on which to apply an generated filters

:return_forbidden_error? (boolean/0) - Whether or not to return a forbidden error in cases of not being authorized. The default value is false.

can_upload_file_to_setting?(actor, key, value, file, params_or_opts \\ %{}, opts \\ [])

Runs authorization checks for CaseManager.Configuration.Setting.upload_file, returning a boolean. See Ash.can?/3 for more information

Options

:maybe_is (term/0) - If the actor may be able to perform the action, what value should be returned. The default value is true.

:filter_with - If set to :error, the query will raise an error on a match. If set to :filter the query will filter out unauthorized access. Valid values are :filter, :error The default value is :filter.

:validate? (boolean/0) - Whether or not to treat an invalid action as a non-allowed action. The default value is false.

:reuse_values? (boolean/0) - Whether or not loaded data like aggregates, calculations and relationships should be checked in memory if possible, instead of querying. No effect if pre_flight? is false. The default value is false.

:pre_flight? (boolean/0) - Whether or not this is a pre_flight check (which may perform optimized in-memory checks) or the final proper check. The default value is true.

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol. Will overwrite any actor, tenant or context provided. See Ash.Context for more.

:run_queries? (boolean/0) - Whether or not to run queries. If set to true, :maybe will not be returned. The default value is true.

:data - The record or records specifically attempting to be acted upon.

:tenant (value that implements the Ash.ToTenant protocol) - The tenant to use for authorization

:alter_source? (boolean/0) - If set to true, the source being authorized is returned so it can be run. The default value is false.

:base_query (term/0) - A base query on which to apply an generated filters

:no_check? (boolean/0) - Whether or not authorization must pass at the strict/filter step, or if post-checks are allowed to be run The default value is false.

:on_must_pass_strict_check (term/0) - Override the value returned when no_check? is true but a check must be run.

:atomic_changeset (term/0) - A base query on which to apply an generated filters

:return_forbidden_error? (boolean/0) - Whether or not to return a forbidden error in cases of not being authorized. The default value is false.

changeset_to_add_setting(key, value, params_or_opts \\ %{}, opts \\ [])

Returns the changeset corresponding to the action.

Options

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user. Valid values are true, false, nil

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:changeset (term/0) - A changeset to seed the action with.

changeset_to_upload_file_to_setting(key, value, file, params_or_opts \\ %{}, opts \\ [])

Returns the changeset corresponding to the action.

Options

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user. Valid values are true, false, nil

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:changeset (term/0) - A changeset to seed the action with.

default_short_name()

upload_file_to_setting(key, value, file, params \\ nil, opts \\ nil)

Calls the upload_file action on CaseManager.Configuration.Setting.

Arguments

key

value

file

Options

:upsert? (boolean/0) - If a conflict is found based on the primary key, the record is updated in the database (requires upsert support) The default value is false.

:return_skipped_upsert? (boolean/0) - If true, and a record was not upserted because its filter prevented the upsert, the original record (which was not upserted) will be returned. The default value is false.

:upsert_identity (atom/0) - The identity to use when detecting conflicts for upsert?, e.g. upsert_identity: :full_name. By default, the primary key is used. Has no effect if upsert?: true is not provided

:upsert_fields - The fields to upsert. If not set, the action's upsert_fields is used, and if that is not set, then any fields not being set to defaults are written.

:upsert_condition (term/0) - An expression to check if the record should be updated when there's a conflict.

:timeout (timeout/0) - A positive integer, or :infinity. If none is provided, the timeout configured on the domain is used.

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:action (term/0) - The action to use, either an Action struct or the name of the action

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user. Valid values are true, false, nil

:context (map/0) - Context to set on the query, changeset, or input

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:return_notifications? (boolean/0) - Use this if you're running ash actions in your own transaction and you want to manually handle sending notifications.
If a transaction is ongoing, and this is false, notifications will be discarded, otherwise the return value is {:ok, result, notifications} (or {:ok, notifications})
To send notifications later, use Ash.Notifier.notify(notifications). It sends any notifications that can be sent, and returns the rest. The default value is false.

:rollback_on_error? (boolean/0) - Whether or not to rollback the transaction on error, if the resource is in a transaction.
If the action has transaction? false this option has no effect. If an error is returned from the data layer and the resource is in a transaction, the transaction is always rolled back, regardless. The default value is true.

:notification_metadata (term/0) - Metadata to be merged into the metadata field for all notifications sent from this operation. The default value is %{}.

:skip_unknown_inputs - A list of inputs that, if provided, will be ignored if they are not recognized by the action. Use :* to indicate all unknown keys.

:load (term/0) - A load statement to add onto the changeset

:changeset (term/0) - A changeset to seed the action with.

:bulk_options (keyword/0) - Options passed to Ash.bulk_create, if a list or stream of inputs is provided.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:after_action (function of arity 2) - An after_action hook to be added to each processed changeset

:read_action (atom/0) - The action to use when building the read query.

:assume_casted? (boolean/0) - Whether or not to cast attributes and arguments as input. This is an optimization for cases where the input is already casted and/or not in need of casting The default value is false.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:authorize_query_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_changeset_with - If set to :error, instead of filtering unauthorized changes, unauthorized changes will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. The default value is :filter.

:sorted? (boolean/0) - Whether or not to sort results by their input position, in cases where return_records?: true was provided. The default value is false.

:return_records? (boolean/0) - Whether or not to return all of the records that were inserted. Defaults to false to account for large inserts. The default value is false.

:return_errors? (boolean/0) - Whether or not to return all of the errors that occur. Defaults to false to account for large inserts. The default value is false.

:batch_size (pos_integer/0) - The number of records to include in each batch. Defaults to the default_limit or max_page_size of the action, or 100.

:return_stream? (boolean/0) - If set to true, instead of an Ash.BulkResult, a mixed stream is returned. Potential elements: {:notification, notification} - if return_notifications? is set to true {:ok, record} - if return_records? is set to true {:error, error} - an error that occurred. May be changeset or an individual error. The default value is false.

:return_nothing? (boolean/0) - Mutes warnings about returning nothing. Only relevant if return_stream? is set to true and all other return_*? options are set to false. The default value is false.

:stop_on_error? (boolean/0) - If true, the first encountered error will stop the action and be returned. Otherwise, errors will be skipped. The default value is false.

:notify? (boolean/0) - Whether or not to generate any notifications. If this is set to true then the data layer must return the results from each batch. This may be intensive for large bulk actions. Notifications will be automatically sent unless return_notifications? is set to true. The default value is false.

:transaction - Whether or not to wrap the entire execution in a transaction, each batch, or not at all. Keep in mind: before_transaction and after_transaction hooks attached to changesets will have to be run inside the transaction if you choose transaction: :all. The default value is :batch.

:max_concurrency (non_neg_integer/0) - If set to a value greater than 0, up to that many tasks will be started to run batches asynchronously The default value is 0.

upload_file_to_setting!(key, value, file, params \\ nil, opts \\ nil)

Calls the upload_file action on CaseManager.Configuration.Setting. Raises any errors instead of returning them

Arguments

key

value

file

Options

:upsert? (boolean/0) - If a conflict is found based on the primary key, the record is updated in the database (requires upsert support) The default value is false.

:return_skipped_upsert? (boolean/0) - If true, and a record was not upserted because its filter prevented the upsert, the original record (which was not upserted) will be returned. The default value is false.

:upsert_identity (atom/0) - The identity to use when detecting conflicts for upsert?, e.g. upsert_identity: :full_name. By default, the primary key is used. Has no effect if upsert?: true is not provided

:upsert_fields - The fields to upsert. If not set, the action's upsert_fields is used, and if that is not set, then any fields not being set to defaults are written.

:upsert_condition (term/0) - An expression to check if the record should be updated when there's a conflict.

:timeout (timeout/0) - A positive integer, or :infinity. If none is provided, the timeout configured on the domain is used.

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:action (term/0) - The action to use, either an Action struct or the name of the action

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user. Valid values are true, false, nil

:context (map/0) - Context to set on the query, changeset, or input

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:return_notifications? (boolean/0) - Use this if you're running ash actions in your own transaction and you want to manually handle sending notifications.
If a transaction is ongoing, and this is false, notifications will be discarded, otherwise the return value is {:ok, result, notifications} (or {:ok, notifications})
To send notifications later, use Ash.Notifier.notify(notifications). It sends any notifications that can be sent, and returns the rest. The default value is false.

:rollback_on_error? (boolean/0) - Whether or not to rollback the transaction on error, if the resource is in a transaction.
If the action has transaction? false this option has no effect. If an error is returned from the data layer and the resource is in a transaction, the transaction is always rolled back, regardless. The default value is true.

:notification_metadata (term/0) - Metadata to be merged into the metadata field for all notifications sent from this operation. The default value is %{}.

:skip_unknown_inputs - A list of inputs that, if provided, will be ignored if they are not recognized by the action. Use :* to indicate all unknown keys.

:load (term/0) - A load statement to add onto the changeset

:changeset (term/0) - A changeset to seed the action with.

:bulk_options (keyword/0) - Options passed to Ash.bulk_create, if a list or stream of inputs is provided.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:after_action (function of arity 2) - An after_action hook to be added to each processed changeset

:read_action (atom/0) - The action to use when building the read query.

:assume_casted? (boolean/0) - Whether or not to cast attributes and arguments as input. This is an optimization for cases where the input is already casted and/or not in need of casting The default value is false.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:authorize_query_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_changeset_with - If set to :error, instead of filtering unauthorized changes, unauthorized changes will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. The default value is :filter.

:sorted? (boolean/0) - Whether or not to sort results by their input position, in cases where return_records?: true was provided. The default value is false.

:return_records? (boolean/0) - Whether or not to return all of the records that were inserted. Defaults to false to account for large inserts. The default value is false.

:return_errors? (boolean/0) - Whether or not to return all of the errors that occur. Defaults to false to account for large inserts. The default value is false.

:batch_size (pos_integer/0) - The number of records to include in each batch. Defaults to the default_limit or max_page_size of the action, or 100.

:return_stream? (boolean/0) - If set to true, instead of an Ash.BulkResult, a mixed stream is returned. Potential elements: {:notification, notification} - if return_notifications? is set to true {:ok, record} - if return_records? is set to true {:error, error} - an error that occurred. May be changeset or an individual error. The default value is false.

:return_nothing? (boolean/0) - Mutes warnings about returning nothing. Only relevant if return_stream? is set to true and all other return_*? options are set to false. The default value is false.

:stop_on_error? (boolean/0) - If true, the first encountered error will stop the action and be returned. Otherwise, errors will be skipped. The default value is false.

:notify? (boolean/0) - Whether or not to generate any notifications. If this is set to true then the data layer must return the results from each batch. This may be intensive for large bulk actions. Notifications will be automatically sent unless return_notifications? is set to true. The default value is false.

:transaction - Whether or not to wrap the entire execution in a transaction, each batch, or not at all. Keep in mind: before_transaction and after_transaction hooks attached to changesets will have to be run inside the transaction if you choose transaction: :all. The default value is :batch.

:max_concurrency (non_neg_integer/0) - If set to a value greater than 0, up to that many tasks will be started to run batches asynchronously The default value is 0.