CaseManagerWeb.Button (Case Manager v0.1.0)

Provides a custom button UI component.

Functions

button(assigns)

Renders a custom button with content.

Custom buttons come with a hero icon if the icon_name atribute is specified.

The text of the button can either be specified using the default slot or the txt attribute.

Custom buttons come in three colors – primary, secondary, and critical. By default, the primary color is used, but the color may be applied by using the color parameter.

Examples

<.button txt="Simple primary text button" />
<.button icon_name="hero-user-plus">Simple primary button with icon</.button>
<.button color="secondary" txt="Simple secondary text button" />
<.button color="critical" txt="Simple critical text button" />
<.button disabled txt="Simple disabled text button" phx-click="show_modal" />
<.button color="disabled" txt="Simple disabled text button" phx-click="show_modal" />

Attributes

colour (:atom) - Defaults to :primary. Must be one of :primary, :secondary, :tertiary, :disabled, or :critical.

icon_name (:string) - name of hero icon used lhs. Defaults to nil.

label (:string) - text written on button. Defaults to nil.

type (:string) - Defaults to nil.

class (:string) - Defaults to nil.

Global attributes are accepted. Supports all globals plus: ["disabled"].

Slots

inner_block