CaseManagerWeb.Helpers (Case Manager v0.1.0)

Module that defines helper function for the web-based part of the application.

Functions

load_bg()

Retrieves the background image from the database. If the image isn't set in the datbase this function returns nil. This is to allow to use :if={@image} to check if this settings has been set.

load_logo()

Retrieves the logo from the database. If the logo isn't set in the datbase this function returns nil. This is to allow to use :if={@logo} to check if this settings has been set.

render_markdown!(text)

Render basic HTML from text with markdown. All HTML tags are stripped before returning it as a raw value.