CaseManagerWeb.IconBtn (Case Manager v0.1.0)

Provides a custom button that only displays an icon.

Functions

icon_btn(assigns)

Renders a square icon button.

Icon buttons come in three colors – primary, secondary, and critical. By default, the primary color is used, but the color may be applied by using the color parameter.

Icon buttons come in three sizes – large and small. By default, the large size is used, but the size may be applied by using the size parameter.

Examples

<.icon_btn icon_name="hero-pause-circle" colour="critical" />
<.icon_btn icon_name="hero-arrow-top-right-on-square" colour="secondary" size="small" class="pl-0.5 pb-1" />

Attributes

size (:atom) - Defaults to :large. Must be one of :large, or :small.

colour (:atom) - Defaults to :primary. Must be one of :primary, :secondary, :tertiary, :disabled, or :critical.

icon_name (:string) (required)

type (:string) - Defaults to nil.

class (:string) - Defaults to nil.

Global attributes are accepted. Supports all globals plus: ["disabled"].