CaseManagerWeb.Input (Case Manager v0.1.0)

Provides insput components with label and error messages.

Functions

input(assigns)

Renders an input with label and error messages.

A Phoenix.HTML.FormField may be passed as argument, which is used to retrieve the input name, id, and values. Otherwise all attributes may be passed explicitly.

Types

This function accepts all HTML input types, considering that:

You may also set type="select" to render a <select> tag

type="checkbox" is used exclusively to render boolean values

For live file uploads, see Phoenix.Component.live_file_input/1

See https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input for more information.

Examples

<.input field={@form[:email]} type="email" />
<.input name="my-input" errors={["oh no!"]} />
<.input type="checkbox" name="checkbox_name" label="Text to checkbox" />

Attributes

id (:any) - Defaults to nil.

name (:any)

label (:string) - Defaults to nil.

value (:any)

type (:string) - Defaults to "text". Must be one of "checkbox", "color", "date", "datetime-local", "email", "file", "hidden", "month", "number", "password", "range", "radio", "search", "select", "tel", "text", "textarea", "time", "url", or "week".

field (Phoenix.HTML.FormField) - a form field struct retrieved from the form, for example: @form[:email].

errors (:list) - Defaults to [].

checked (:boolean) - the checked flag for checkbox inputs.

prompt (:string) - the prompt for select inputs. Defaults to nil.

options (:list) - the options to pass to Phoenix.HTML.Form.options_for_select/2.

multiple (:boolean) - the multiple flag for select inputs. Defaults to false.

class (:string) - manually assigned classes. Defaults to nil.

Global attributes are accepted. Supports all globals plus: ["accept", "autocomplete", "capture", "cols", "disabled", "form", "list", "max", "maxlength", "min", "minlength", "multiple", "pattern", "placeholder", "readonly", "required", "rows", "size", "step"].

Slots

inner_block