CaseManagerWeb.Table (Case Manager v0.1.0)

Provides a table.

Functions

table(assigns)

Renders a table with generic styling.

Examples

<.table id="users" rows={@users}>
  <:col :let={user} label="id"><%= user.id %></:col>
  <:col :let={user} label="username"><%= user.username %></:col>
</.table>

Attributes

id (:string) (required)

rows (:list) (required)

row_id (:any) - the function for generating the row id. Defaults to nil.

row_click (:any) - the function for handling phx-click on each row. Defaults to nil.

row_item (:any) - the function for mapping each row before calling the :col and :action slots. Defaults to &Function.identity/1.

Slots

col (required) - Accepts attributes:

label (:any)

width (:string) - optionally specify a width if a fixed width is wanted. Otherwise it will space the column out automatically using the space available.

not_clickable_area? (:boolean) - true if col contains a clickable item.

action - the slot for showing user actions in the last table column.